+++ CHANGES.md
@@ -13,6 +13,7 @@
 + read `GARGS_PROCESS_BUFFER` to let user set size of data before a tempfile is used.
 + read `GARGS_WAIT_MULTIPLIER` to determine how many finished processes will wait for single slow processes
   higher values improve concurrency at the expense of memory.
++ better cleanup of tmp files in case of process halt.
 
 0.3.4
 =====